# src/generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BookStatus": {
				"type": "string",
				"enum": [
					"New",
					"Used",
					"Out Of Stock"
				]
			},
			"Book": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/BookStatus"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"author",
					"title",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Book_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/BookStatus"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-starter",
		"version": "1.0.0",
		"description": "Starter Kit for Express Api with Typescript, Unit and Integration Testing",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Gary Ivy"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/books": {
			"post": {
				"operationId": "PostBook",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"description": "Adds a book",
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Book"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Book"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page_size",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "page_number",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					}
				]
			}
		},
		"/books/{bookId}": {
			"put": {
				"operationId": "PutBook",
				"responses": {
					"200": {
						"description": "Updated"
					},
					"404": {
						"description": "Unable to find book"
					}
				},
				"description": "Updates one or more book properties",
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Book_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Book"
								}
							}
						}
					},
					"404": {
						"description": "Unable to find book"
					}
				},
				"description": "Get a book by ID",
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/health/readiness": {
			"get": {
				"operationId": "GetReadiness",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Probe for determine if application is ready to receive requests",
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/health/liveness": {
			"get": {
				"operationId": "GetLiveness",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Probe for determine if application is can still receive requests",
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}